% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collide.R
\name{col_detect_broad}
\alias{col_detect_broad}
\title{Detect bounding box collision}
\usage{
col_detect_broad(nr, x, y, coldf, hjust = 0, vjust = 0)
}
\arguments{
\item{nr}{primary native raster which will be tested against all the 
native rasters represented in \code{coldf}}

\item{x, y}{Scalar coordinates of \code{nr}}

\item{coldf}{collision data.frame returned by \code{col_setup()}}

\item{hjust, vjust}{specify horizontal and vertical justification of the 
raster.  e.g. \code{hjust = vjust = 0} the blitting
starts at the top-left of the image. Use \code{hjust = vjust = 0.5}
to treat the centre of the \code{src_} as the blitting origin.
Default (0, 0)}
}
\value{
logical vector with an entry for each row in \code{coldf}; TRUE
        when bounding box for \code{nr} overlaps any of the bounding boxes for
        the native rasters in \code{coldf}, otherwise FALSE.
}
\description{
Detect bounding box collision
}
