% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collide.R
\name{col_setup}
\alias{col_setup}
\title{Setup the collision structure}
\usage{
col_setup(nr_list, x, y, hjust = 0, vjust = 0)
}
\arguments{
\item{nr_list}{list of native rasters}

\item{x, y}{coordinates for each native raster}

\item{hjust, vjust}{specify horizontal and vertical justification of the 
raster.  e.g. \code{hjust = vjust = 0} the blitting
starts at the top-left of the image. Use \code{hjust = vjust = 0.5}
to treat the centre of the \code{src_} as the blitting origin.
Default (0, 0)}
}
\value{
collision data.frame with stored information about all the given 
        native rasters.  Each row represents the corresponding object in the
        \code{nr_list}
}
\description{
Setup the collision structure
}
\examples{
nr1 <- nara::nr_new(5, 2, 'white')
nr2 <- nara::nr_new(4, 4, 'grey')
nr_list <- list(nr1, nr2)
coldf <- col_setup(nr_list, x = c(0, 10), y = c(100, 10))
coldf
}
