% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collide.R
\name{col_detect_narrow}
\alias{col_detect_narrow}
\title{Detect overlapping non-transparent pixels}
\usage{
col_detect_narrow(nr, x, y, coldf)
}
\arguments{
\item{nr}{primary native raster which will be tested against all the 
native rasters represented in \code{coldf}}

\item{x, y}{Scalar coordinates of \code{nr}}

\item{coldf}{collision data.frame returned by \code{col_setup()}}
}
\value{
logical vector with an entry for each row in \code{coldf}; TRUE
        when any non-transparent pixel in \code{nr} overlaps any 
        non-transparent pixel in 
        the native rasters in \code{coldf}, otherwise FALSE.
}
\description{
Detect overlapping non-transparent pixels
}
\examples{
nr1 <- nara::nr_new(5, 2, 'white')
nr2 <- nara::nr_new(4, 4, 'grey')
nr_list <- list(nr1, nr2)
coldf <- col_setup(nr_list, x = c(0, 10), y = c(100, 10))
coldf 

}
